Feat: Split timeline DTOs into namespaces
Added

- Timeline and trace event DTOs under the Timelines namespace
- TraceEvent copy helpers for branch and pull request state changes
- Test imports updated for the new DTO namespace layout

Changed

- DTO files moved out of the flat GIt.Dto namespace into Timelines
- TraceEvent now exposes branch and pull request copy helpers used by DTO conversions
- Serialization tests now reference the new DTO namespace split

Result

Timeline DTOs now live in a folder layout that matches their namespace and future DTO growth stays isolated from the rest of the GIt surface.
